moved get_variables_metadata.py to data_request_api.command_line, added pyproject.toml entry point
