+++ pyproject.toml
@@ -38,6 +38,7 @@ Issues = "https://github.com/CMIP-Data-Request/CMIP7_DReq_Software/issues"
 # modules and functions that act as command line utilities to be specified here
 # use pip install -e . from the root directory of the repository to set up the entry points
 export_dreq_lists_json = "data_request_api.command_line.export_dreq_lists_json:main"
+get_variables_metadata = "data_request_api.command_line.get_variables_metadata:main"
 
 [tool.pytest.ini_options]
 cache_dir = "tests/.pytest_cache"
