+++ scripts/workflow_example.py
@@ -29,17 +29,16 @@ Usage examples:
     ./export_dreq_lists_json.py v1.0 dreq_list.json --opportunities_file opps.json
 
 '''
-import sys
 import os
+import sys
 
+# add repo top-level dir to system path so that data_request_api imports work from scripts/
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
 
 from data_request_api.stable.content.dreq_api import dreq_content as dc
 from data_request_api.stable.query import dreq_query as dq
 from importlib import reload
 reload(dq)
-
 
 use_dreq_version = 'v1.1'
